requirements.txt
# Core Dependencies
tensorflow>=2.13.0,<2.16.0
streamlit>=1.28.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0

# Image Processing
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0

# Optional: For better performance
# tensorflow-gpu>=2.13.0  # Uncomment if you have CUDA-enabled GPU

# Development Dependencies (Optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# pylint>=2.17.0